[refactor] Move interpreter code
